requirements.txt
Flask==2.3.3
pymongo==4.5.0
python-dotenv==1.0.0
flask-mail==0.9.1
werkzeug==2.3.7
jinja2==3.1.2
email-validator==2.0.0
